% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19}
\alias{covid19}
\title{COVID19 in Europe}
\format{
\subsection{\code{covid19}}{

A data frame with 28729 rows and 11 columns:
\describe{
\item{dateRep}{Date}
\item{day, month, year}{Date indicators}
\item{cases}{Documented COVID19 cases}
\item{deaths}{COVID19 related deaths}
\item{countriesAndTerritories}{Country name}
\item{geoId, countryterritoryCode}{Codes to indicate country}
\item{popData2020}{Eurostat 2020 data}
\item{continentExp}{Continent}
}
}
}
\source{
\url{https://opendata.ecdc.europa.eu/covid19/nationalcasedeath_eueea_daily_ei/csv}
}
\usage{
covid19
}
\description{
A subset of data from the ECDC about COVID19 cases and related deaths in different European countries
Report .
}
\keyword{datasets}
