% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_cap.R
\name{graph_cap}
\alias{graph_cap}
\title{graph_cap}
\usage{
graph_cap(content)
}
\arguments{
\item{content}{Caption content}
}
\value{
Caption that fits well in the graph
}
\description{
graph_cap
}
\examples{
graph_cap("Figure 1. The caption can include any string of any length and will be shortened to a length suitable for the size of a graph.")
}
