% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_jitter.R
\name{graph_jitter}
\alias{graph_jitter}
\title{graph_jitter}
\usage{
graph_jitter(numb)
}
\arguments{
\item{numb}{Amount of jitter}
}
\value{
Standard input for geom_point's position
}
\description{
graph_jitter
}
