% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinesin}
\alias{kinesin}
\title{PDB file containing the information about 3COB (kinesin)
PDB 3cob, kinesin}
\format{
PDB format
A list with 28729 rows and 11 columns:
}
\source{
\url{https://www.rcsb.org/structure/3COB}
}
\usage{
kinesin
}
\description{
PDB file containing the information about 3COB (kinesin)
PDB 3cob, kinesin
}
\keyword{datasets}
