% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_vis.R
\name{prot_vis}
\alias{prot_vis}
\title{prot_vis}
\usage{
prot_vis(mydata, style)
}
\arguments{
\item{mydata}{File path to pdb file}

\item{style}{"Cartoon", "spheres" or "sticks"}
}
\value{
3D structural visualisation of the protein
}
\description{
prot_vis
}
\examples{
prot_vis(kinesin, "cartoon")
}
