% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_cartoon.R
\name{protein_cartoon}
\alias{protein_cartoon}
\title{protein_cartoon}
\usage{
protein_cartoon(mydata)
}
\arguments{
\item{mydata}{File path to pdb file}
}
\value{
3D structural visualization of the protein in cartoon style
}
\description{
protein_cartoon
}
\examples{
protein_cartoon(kinesin)
}
