% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_sphere.R
\name{protein_sphere}
\alias{protein_sphere}
\title{protein_sphere}
\usage{
protein_sphere(mydata)
}
\arguments{
\item{mydata}{R object of PDB file made using readLines}
}
\value{
3D structural visualization of the protein in sphere style
}
\description{
protein_sphere
}
\examples{
protein_sphere(kinesin)
}
