% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_spheres.R
\name{protein_spheres}
\alias{protein_spheres}
\title{protein_spheres}
\usage{
protein_spheres(mydata)
}
\arguments{
\item{mydata}{File path to pdb file}
}
\value{
3D structural visualization of the protein in sphere style
}
\description{
protein_spheres
}
\examples{
protein_spheres(kinesin)
}
