% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_stick.R
\name{protein_stick}
\alias{protein_stick}
\title{protein_stick}
\usage{
protein_stick(mydata)
}
\arguments{
\item{mydata}{R object of PDB file made using readLines}
}
\value{
3D structural visualization of the protein in stick style
}
\description{
protein_stick
}
\examples{
protein_stick(kinesin)
}
