% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_sticks.R
\name{protein_sticks}
\alias{protein_sticks}
\title{protein_sticks}
\usage{
protein_sticks(mydata)
}
\arguments{
\item{mydata}{File path to pdb file}
}
\value{
3D structural visualization of the protein in stick style
}
\description{
protein_sticks
}
\examples{
protein_sticks(kinesin)
}
