% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_visual.R
\name{protein_visual}
\alias{protein_visual}
\title{protein_visual}
\usage{
protein_visual(pdbobject)
}
\arguments{
\item{pdbobject}{R object of PDB file made using readLines}
}
\value{
3D structural visualization of the protein in sticks and spherical style with PyMOL colouring
}
\description{
protein_visual
}
\examples{
protein_visual(kinesin)
}
