% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_csv_rds.R
\name{save_csv_rds}
\alias{save_csv_rds}
\title{save_csv_rds}
\usage{
save_csv_rds(df, output, rownames = FALSE)
}
\arguments{
\item{df}{dataframe to save as csv and rds}

\item{output}{output location including filename}

\item{rownames}{rownames or not}
}
\value{
A csv and rds file
}
\description{
save_csv_rds
}
\examples{
save_csv_rds(covid19, "covid19")
}
