% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_minmax.R
\name{state_minmax}
\alias{state_minmax}
\title{state_minmax}
\usage{
state_minmax(col)
}
\arguments{
\item{col}{Column to state the min and max value from}
}
\value{
String with "From min to max"
}
\description{
state_minmax
}
