% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_cartoon.R
\name{protein_cartoon}
\alias{protein_cartoon}
\title{protein_cartoon}
\usage{
protein_cartoon(pdbobject)
}
\arguments{
\item{pdbobject}{R object of PDB file made using readLines}
}
\value{
3D structural visualization of the protein in cartoon style with PyMOL colouring
}
\description{
protein_cartoon
}
\examples{
protein_cartoon(kinesin)
}
