% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_csv_rds.R
\name{save_csv_rds}
\alias{save_csv_rds}
\title{Title}
\usage{
save_csv_rds(df, output, rownames = FALSE)
}
\arguments{
\item{df}{dataframe to save as csv and rds}

\item{output}{output location including filename}

\item{rownames}{rownames or not}
}
\value{
A csv and rds file
}
\description{
Title
}
\examples{
mydata <- data.frame(col1 = c("val1", "val2", "val3"), col2 = c("val1", "val2", "val3"))
save_csv_rds(mydata, "mydata")
}
